Move taskmaster and executor logic to separate module
