+++ cluster-manager/standalone/src/test/scala/clustermanager/standalone/taskmanager/utils/HardwareSpec.scala
@@ -1,6 +1,7 @@
 package clustermanager.standalone.taskmanager.utils
 
 import runtime.common.BaseSpec
+import runtime.taskmaster.common.Hardware
 
 
 class HardwareSpec extends BaseSpec {
