+++ runtime/appmanager/src/main/resources/appmanager.conf
@@ -55,7 +55,7 @@ akka {
 }
 
 appmanager {
-  clustermanager.standalone.resourcemanager = "yarn"
+  resourcemanager = ""
   appMasterKeepAlive = 500
   restPort = 5050
   interface = "localhost"
