+++ runtime/taskmaster/src/main/scala/runtime/taskmaster/common/Config.scala
@@ -1,0 +1,5 @@
+package runtime.taskmaster.common
+
+final case class TaskMasterConf(heartbeat: Long, hostname: String)
+final case class TaskExecutorConf(monitorTicker: Long)
+
